requirement.txt
asgiref==3.8.1
Django==5.1
django-autoslug==1.9.9
pillow==10.4.0
sqlparse==0.5.1
